Setup: added webpack and npm
Added webpack along with a development server and npm as a package
manager. Also added npm commands to run the server and build the
project. To run the server, type "npm run serve", and type "npm run
build" to build.

HTML and CSS files still need to be in the dist folder, but
should eventually be placed within ./src. For this to work, webpack will
most likely need another loader. Also note that things such as a
formater and a linter such as ESlint is missing.