Bugfix: fixed problems with Webpack
Removed onchange attribute for element with id "font-size" and added an
event listener instead. Note that the Event object passed to the
callback function is not the same when using event listeners as it is
with the HTML onchange attribute. It requires the use of the method
.target() before calling a method such as .value(). Also moved code from
script.ts into index.ts.

In the HTML template (the file within src), "require()" was added to get
webpack to recognize the .css file as an asset.