Refactor: separated code running in the render process
Separated code running in the render process into a folder called
'renderer'. This is done to avoid calling scripts that are intended to
be a part of the main process. A render process (a HTML document in this case)
can not execute scripts that use require() or any other forms of imports.