Setup: added Electron
* Setup: added electron

Note that file paths are messed up, flickering happens on at least one
device (the one this message was written on) and imports do not work
(related to file paths). Mathitor files also need to be bundled with the
application (this was needed previously too).

* Setup: fixed electron (sort of)

* Refactor: separated code running in the render process

Separated code running in the render process into a folder called
'renderer'. This is done to avoid calling scripts that are intended to
be a part of the main process. A render process (a HTML document in this case)
can not execute scripts that use require() or any other forms of imports.

* Other: fixed MathQuill

* Other: updated dependencies for Electron to run without npx