Feature: added incomplete functionality of autocomplete
The feature still needs a way of selecting a suggested completion, a
complete UI and tests need to be written. It only supports
completing the right-most text. A letter also needs to be written for it
to show up, which may or may not be an issue. Also, when focusing on a
mathfield with text already written, the autocompletion tab is not
shown.