Feature: added suggestion tab (#31)
* Feature: added class for autocomplete element

* Bugfix: fixed some autoCommand stuff

* Feature: added basic autocompletion logic

Added basic, very inefficient logic for autocompletion. Autocompletion
still requires a usable interface and knowledge about where the caret is
positioned to give suggestions when editing previously written text.
Also fixed a problem with autoCommands and autoOperators.

* Feature: added incomplete functionality of autocomplete

The feature still needs a way of selecting a suggested completion, a
complete UI and tests need to be written. It only supports
completing the right-most text. A letter also needs to be written for it
to show up, which may or may not be an issue. Also, when focusing on a
mathfield with text already written, the autocompletion tab is not
shown.

* Feature: added TODOs

* Refactor: enclosed Autocomplete within class

* Refactor: removed unnecessary method arguments

* Feature: added movement to suggestion tab

* Feature: added "scrolling"

* Refactor: moved methods around

* Feature: added ability to move from and to the tab

* Feature: added ability to complete with suggestion

* Other: added TODO for bug

* Feature: improved suggestions

Suggestions now take into account where the caret is.

* Bugfix: fixed bug when entering tab

* Refactor: updated sentinel value for marking

Replaced the old emoji based sentinel value with noncharacter-FFFF. This
is to lessen the already extremly low risk of user input being the
sentinel character used for testing where the marker is.

* Documentation: documented methods

Refactored some code and added documentation to most methods. Some
methods miss documentation due to it being clear enough what they do
based to their name.

* Refactor: moved initializations

* Feature: added configuration for suggestions

Moved css and json files into new folder "assets", added ipc functions
to load default suggestions. Also added functions to load user defined
suggestions, but these are not used as of yet.

* Feature: added default suggestions