Bugfix: resolved issue regarding math fields
Previously, clicking on another math field while already being in one
would not work. This was due to a method called by an event listener.
This has now been resolved and the event listener has now been added to
the window instead of to all math fields. It now calls a function to
close the suggestion tab. The previously intended behaviour was for the
suggestion tab to be updated with new suggestions. The current behaviour
is that the suggestion tab is closed when the user clicks anywhere.